pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.0","poetry-dynamic-versioning>=0.22.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
bump = true

[tool.poetry-dynamic-versioning.substitution]
files = ["src/ndn/__init__.py"]

[tool.poetry]
name = "python-ndn"
version = "0.0.0"
description = "An NDN client library with AsyncIO support in Python 3"
authors = ["Xinyu Ma <ma.xinyu.26a@kyoto-u.jp>"]
license = "Apache-2.0"
readme = ["README.rst", "CHANGELOG.rst"]
homepage = "https://python-ndn.readthedocs.io"
repository = "https://github.com/named-data/python-ndn"
documentation = "https://python-ndn.readthedocs.io"
keywords = ["NDN"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet",
    "Topic :: System :: Networking",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{ include = "ndn", from = "src" }]
include = [{ path = "tests", format = "sdist" }]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/named-data/python-ndn/issues"
"Changelog" = "https://python-ndn.readthedocs.io/en/latest/src/changelog.html"

[tool.poetry.dependencies]
python = "^3.10"
pycryptodomex = "^3.18.0"
pygtrie = "^2.5.0"
aenum = "^3.1.12"
lark = "^1.1.7"
aiohttp = "^3.8.5"

# Extra dependencies [dev]
pytest = { version = "^7.1.2", optional = true }
pytest-cov = { version = "^4.1.0", optional = true }
flake8 = { version = "^6.1.0", optional = true }

# Extra dependencies [docs]
Sphinx = { version = "^7.1.2", optional = true }
sphinx-rtd-theme = { version = "^1.2.2", optional = true }
sphinx-autodoc-typehints = { version = "^1.24.0", optional = true }

[tool.poetry.extras]
dev = ["pytest", "pytest-cov", "flake8"]
docs = ["Sphinx", "sphinx-rtd-theme", "sphinx-autodoc-typehints"]

[tool.poetry.scripts]
pyndnsec = "ndn.bin.sec:main"
pyndntools = "ndn.bin.tools:main"
pynfdc = "ndn.bin.nfdc:main"
